Refactor imports in SavedRoutesComponent and SelectedItineraryContext
